+++ app/services/submit_support_request.rb
@@ -89,14 +89,17 @@ private
 
   # @return [Support::Case] TODO: Move into inbound API
   def open_case
+    org = Support::Organisation.find_by(urn: request.school_urn)
+
     @kase = Support::Case.create!(request_text: request.message_body,
                                   first_name: user.first_name,
                                   last_name: user.last_name,
                                   email: user.email,
                                   phone_number: request.phone_number,
-                                  organisation_urn: request.school_urn,
-                                  organisation_name: request.school_name,
+                                  organisation: org,
                                   category: map_category)
+
+    record_case_opening
 
     Support::Interaction.create!({  case: @kase,
                                     event_type: 4,
@@ -111,6 +114,13 @@ private
 
     @kase.documents << document if request.journey
     @kase
+  end
+
+  def record_case_opening
+    Support::RecordAction.new(
+      case_id: @kase.id,
+      action: "open_case",
+    ).call
   end
 
   # API (draft) ----------------------------------------------------------------