+++ spec/models/self-serve/activity_log_item_spec.rb
@@ -4,4 +4,12 @@ RSpec.describe ActivityLogItem, type: :model do
     it { is_expected.to validate_presence_of(:user_id) }
     it { is_expected.to validate_presence_of(:action) }
   end
+
+  describe "#to_csv" do
+    it "includes headers" do
+      expect(described_class.to_csv).to eql(
+        "id,journey_id,user_id,contentful_category_id,contentful_section_id,contentful_task_id,contentful_step_id,action,data,created_at,updated_at\n",
+      )
+    end
+  end
 end